requirements.txt
pandas
numpy
sqlalchemy
sqlalchemy-utils
apscheduler
scikit-learn
python-bittrex
requests
pytz
